requirements.txt
streamlit==1.31.0
openai==1.13.3
llama-index==0.9.15
pandas==2.1.1
pillow==10.1.0
python-dotenv==1.0.0
toml==0.10.2
matplotlib==3.8.0
plotly==5.18.0
scikit-learn==1.3.2
requests==2.31.0